backend/requirements_modern.txt
Django>=5.1
djangorestframework>=3.15
django-filter>=24.2
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3
PyJWT>=2.8
gunicorn>=21.2
